refactor: pages 내 디렉토리 컨벤션 수정
